Add separate tests for document splitting and English word extraction
