+++ requirements.txt
@@ -1,3 +1,4 @@
 python-docx
 Pillow
 customtkinter
+pytest
